[PandosProgramming] cleanup python code, makes a class
